+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
 - Added Campus Dining subsection to Dining [#5872](https://github.com/rokwire/illinois-app/issues/5872).
+- Show icon for online courses [#5931](https://github.com/rokwire/illinois-app/issues/5931).
 ### Changed
 - Match Notifications buttons to match app style [#5868](https://github.com/rokwire/illinois-app/issues/5868).